Upgrade Besar Besaran

+++ app/Actions/Jetstream/DeleteUser.php
@@ -3,6 +3,7 @@
 namespace App\Actions\Jetstream;
 
 use App\Models\User;
+use App\Support\Permissions;
 use Laravel\Jetstream\Contracts\DeletesUsers;
 
 class DeleteUser implements DeletesUsers
@@ -12,6 +13,8 @@ class DeleteUser implements DeletesUsers
      */
     public function delete(User $user): void
     {
+        if (! $user->can(Permissions::SYSTEM_MANAGE)) abort(403);
+
         $user->deleteProfilePhoto();
         $user->tokens->each->delete();
         $user->delete();
